Create Nested Floats to String
I can easily convert floats to string and string to floats to remedy this but the main problem is that it's not allowing the Enter Key or "" to be input at all so it has nothing to compare. Try: and Except: are the only ways I know how to check a users input as a float or string. Unless I learn something new that isn't covered in the chapters I'm learning currently, I have no clue. Anyone?